Update wbb.go and merge code in GDP PacketIn and PacketOut (#609)
* Add Priority field and update the way to add match field

* move packetin_test.go code and add Priority filed in GDP entry programming

* move packetout_test.go code and add the missing step related to programming entry

* change variable name from Id to ID
