+++ feature/experimental/p4rt/ate_tests/google_discovery_protocol_packetin_test/google_discovery_protocol_packetin_test.go
@@ -19,11 +19,15 @@ import (
 	"errors"
 	"flag"
 	"fmt"
+	"io"
 	"sort"
 	"testing"
+	"time"
 
 	"github.com/cisco-open/go-p4/p4rt_client"
 	"github.com/cisco-open/go-p4/utils"
+	"github.com/google/gopacket"
+	"github.com/google/gopacket/layers"
 	"github.com/openconfig/featureprofiles/feature/experimental/p4rt/wbb"
 	"github.com/openconfig/featureprofiles/internal/attrs"
 	"github.com/openconfig/featureprofiles/internal/deviations"
@@ -45,9 +49,9 @@ var (
 	streamName            = "p4rt"
 	gdpMAC                = "00:0a:da:f0:f0:f0"
 	gdpEtherType          = *ygot.Uint32(0x6007)
-	deviceId              = *ygot.Uint64(1)
-	portId                = *ygot.Uint32(10)
-	electionId            = *ygot.Uint64(100)
+	deviceID              = *ygot.Uint64(1)
+	portID                = *ygot.Uint32(10)
+	electionID            = *ygot.Uint64(100)
 	METADATA_INGRESS_PORT = *ygot.Uint32(1)
 	METADATA_EGRESS_PORT  = *ygot.Uint32(2)
 )
@@ -77,6 +81,175 @@ var (
 		IPv4Len: ipv4PrefixLen,
 	}
 )
+
+type PacketIO interface {
+	GetTableEntry(delete bool) []*wbb.ACLWbbIngressTableEntryInfo
+	GetPacketTemplate() *PacketIOPacket
+	GetTrafficFlow(ate *ondatra.ATEDevice, frameSize uint32, frameRate uint64) []*ondatra.Flow
+	GetEgressPort() []string
+	GetIngressPort() string
+}
+
+type PacketIOPacket struct {
+	SrcMAC, DstMAC *string
+	EthernetType   *uint32
+}
+
+type testArgs struct {
+	ctx      context.Context
+	leader   *p4rt_client.P4RTClient
+	follower *p4rt_client.P4RTClient
+	dut      *ondatra.DUTDevice
+	ate      *ondatra.ATEDevice
+	top      *ondatra.ATETopology
+	packetIO PacketIO
+}
+
+// programmTableEntry programs or deletes p4rt table entry based on delete flag.
+func programmTableEntry(ctx context.Context, t *testing.T, client *p4rt_client.P4RTClient, packetIO PacketIO, delete bool) error {
+	t.Helper()
+	err := client.Write(&p4_v1.WriteRequest{
+		DeviceId:   deviceID,
+		ElectionId: &p4_v1.Uint128{High: uint64(0), Low: electionID},
+		Updates: wbb.ACLWbbIngressTableEntryGet(
+			packetIO.GetTableEntry(delete),
+		),
+		Atomicity: p4_v1.WriteRequest_CONTINUE_ON_ERROR,
+	})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// decodePacket decodes L2 header in the packet and returns destination MAC and ethernet type.
+func decodePacket(t *testing.T, packetData []byte) (string, layers.EthernetType) {
+	t.Helper()
+	packet := gopacket.NewPacket(packetData, layers.LayerTypeEthernet, gopacket.Default)
+	etherHeader := packet.Layer(layers.LayerTypeEthernet)
+	if etherHeader != nil {
+		header, decoded := etherHeader.(*layers.Ethernet)
+		if decoded {
+			return header.DstMAC.String(), header.EthernetType
+		}
+	}
+	return "", layers.EthernetType(0)
+}
+
+// testTraffic sends traffic flow for duration seconds.
+func testTraffic(t *testing.T, ate *ondatra.ATEDevice, flows []*ondatra.Flow, srcEndPoint *ondatra.Interface, duration int) {
+	t.Helper()
+	for _, flow := range flows {
+		flow.WithSrcEndpoints(srcEndPoint).WithDstEndpoints(srcEndPoint)
+	}
+	ate.Traffic().Start(t, flows...)
+	time.Sleep(time.Duration(duration) * time.Second)
+
+	ate.Traffic().Stop(t)
+}
+
+// fetchPackets reads p4rt packets sent to p4rt client.
+func fetchPackets(ctx context.Context, t *testing.T, client *p4rt_client.P4RTClient, expectNumber int) []*p4rt_client.P4RTPacketInfo {
+	t.Helper()
+	packets := []*p4rt_client.P4RTPacketInfo{}
+	for i := 0; i < expectNumber; i++ {
+		_, packet, err := client.StreamChannelGetPacket(&streamName, 0)
+		if err == io.EOF {
+			t.Logf("EOF error is seen in PacketIn.")
+			break
+		} else if err == nil {
+			if packet != nil {
+				packets = append(packets, packet)
+			}
+		} else {
+			t.Fatalf("There is error seen when receving packets. %v, %s", err, err)
+			break
+		}
+	}
+	return packets
+}
+
+// testPacketIn programs p4rt table entry and sends traffic related to GDP,
+// then validates packetin message metadata and payload.
+func testPacketIn(ctx context.Context, t *testing.T, args *testArgs) {
+	leader := args.leader
+	follower := args.follower
+
+	// Insert wbb acl entry on the DUT
+	if err := programmTableEntry(ctx, t, leader, args.packetIO, false); err != nil {
+		t.Fatalf("There is error when programming entry")
+	}
+	// Delete wbb acl entry on the device
+	defer programmTableEntry(ctx, t, leader, args.packetIO, true)
+
+	// Send GDP traffic from ATE
+	srcEndPoint := args.top.Interfaces()[atePort1.Name]
+	testTraffic(t, args.ate, args.packetIO.GetTrafficFlow(args.ate, 300, 2), srcEndPoint, 10)
+
+	packetInTests := []struct {
+		desc       string
+		client     *p4rt_client.P4RTClient
+		expectPass bool
+	}{{
+		desc:       "PacketIn to Primary Controller",
+		client:     leader,
+		expectPass: true,
+	}, {
+		desc:       "PacketIn to Secondary Controller",
+		client:     follower,
+		expectPass: false,
+	}}
+
+	for _, test := range packetInTests {
+		t.Run(test.desc, func(t *testing.T) {
+			// Extract packets from PacketIn message sent to p4rt client
+			packets := fetchPackets(ctx, t, test.client, 40)
+
+			if !test.expectPass {
+				if len(packets) > 0 {
+					t.Fatalf("Unexpected packets received.")
+				}
+			} else {
+				if len(packets) == 0 {
+					t.Fatalf("There are no packets received.")
+				}
+				t.Logf("Start to decode packet and compare with expected packets.")
+				wantPacket := args.packetIO.GetPacketTemplate()
+				for _, packet := range packets {
+					if packet != nil {
+						if wantPacket.DstMAC != nil && wantPacket.EthernetType != nil {
+							dstMac, etherType := decodePacket(t, packet.Pkt.GetPayload())
+							if dstMac != *wantPacket.DstMAC || etherType != layers.EthernetType(*wantPacket.EthernetType) {
+								t.Fatalf("Packet in PacketIn message is not matching wanted packet.")
+							}
+						}
+
+						metaData := packet.Pkt.GetMetadata()
+						for _, data := range metaData {
+							if data.GetMetadataId() == METADATA_INGRESS_PORT {
+								if string(data.GetValue()) != args.packetIO.GetIngressPort() {
+									t.Fatalf("Ingress Port Id is not matching expectation.")
+								}
+							}
+							if data.GetMetadataId() == METADATA_EGRESS_PORT {
+								found := false
+								for _, portData := range args.packetIO.GetEgressPort() {
+									if string(data.GetValue()) == portData {
+										found = true
+									}
+								}
+								if !found {
+									t.Fatalf("Egress Port Id is not matching expectation.")
+								}
+
+							}
+						}
+					}
+				}
+			}
+		})
+	}
+}
 
 func TestMain(m *testing.M) {
 	fptest.RunTests(m)
@@ -151,7 +324,7 @@ func configureDeviceId(ctx context.Context, t *testing.T, dut *ondatra.DUTDevice
 	component := telemetry.Component{}
 	component.IntegratedCircuit = &telemetry.Component_IntegratedCircuit{}
 	component.Name = ygot.String(*p4rtNodeName)
-	component.IntegratedCircuit.NodeId = ygot.Uint64(deviceId)
+	component.IntegratedCircuit.NodeId = ygot.Uint64(deviceID)
 	dut.Config().Component(*p4rtNodeName).Replace(t, &component)
 }
 
@@ -159,7 +332,7 @@ func configureDeviceId(ctx context.Context, t *testing.T, dut *ondatra.DUTDevice
 func configurePortId(ctx context.Context, t *testing.T, dut *ondatra.DUTDevice) {
 	ports := sortPorts(dut.Ports())
 	for i, port := range ports {
-		dut.Config().Interface(port.Name()).Id().Replace(t, uint32(i)+portId)
+		dut.Config().Interface(port.Name()).Id().Replace(t, uint32(i)+portID)
 	}
 }
 
@@ -169,9 +342,9 @@ func setupP4RTClient(ctx context.Context, args *testArgs) error {
 	// Setup p4rt-client stream parameters
 	streamParameter := p4rt_client.P4RTStreamParameters{
 		Name:        streamName,
-		DeviceId:    deviceId,
+		DeviceId:    deviceID,
 		ElectionIdH: uint64(0),
-		ElectionIdL: electionId,
+		ElectionIdL: electionID,
 	}
 
 	// Send ClientArbitration message on both p4rt leader and follower clients.
@@ -206,8 +379,8 @@ func setupP4RTClient(ctx context.Context, args *testArgs) error {
 
 	// Send SetForwardingPipelineConfig for p4rt leader client.
 	if err := args.leader.SetForwardingPipelineConfig(&p4_v1.SetForwardingPipelineConfigRequest{
-		DeviceId:   deviceId,
-		ElectionId: &p4_v1.Uint128{High: uint64(0), Low: electionId},
+		DeviceId:   deviceID,
+		ElectionId: &p4_v1.Uint128{High: uint64(0), Low: electionID},
 		Action:     p4_v1.SetForwardingPipelineConfigRequest_VERIFY_AND_COMMIT,
 		Config: &p4_v1.ForwardingPipelineConfig{
 			P4Info: &p4Info,
@@ -229,7 +402,7 @@ func getGDPParameter(t *testing.T) PacketIO {
 			DstMAC:       &gdpMAC,
 			EthernetType: &gdpEtherType,
 		},
-		IngressPort: fmt.Sprint(portId),
+		IngressPort: fmt.Sprint(portID),
 	}
 }
 
@@ -290,6 +463,7 @@ func (gdp *GDPPacketIO) GetTableEntry(delete bool) []*wbb.ACLWbbIngressTableEntr
 		Type:          actionType,
 		EtherType:     0x6007,
 		EtherTypeMask: 0xFFFF,
+		Priority:      1,
 	}}
 }
